Imaginary part of MathLogGamma was sometimes off by a multiple of 2pi
Fixed an embarassing (and very old) bug in MathLogGamma, which led to
incorrect imaginary part being returned when the imaginary part of
the argument was large. The error was always a multiple of 2pi, so that
as long as the result of the MathLogGamma was used in an exponent, all
final results remained correct. Bullet dogged!